[Silver III] Title: 주유소, Time: 376 ms, Memory: 38020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -1,4 +1,5 @@
 import java.io.*;
+import java.math.BigInteger;
 import java.util.*;
 
 public class Main {
@@ -6,8 +7,7 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
-
-        int price = 0;   // 거리에 따라 업데이트되는 실시간 가격 // 업데이트 안해도 되겠는데..?
+        long price = 0;   // 거리에 따라 업데이트되는 실시간 가격 // 업데이트 안해도 되겠는데..?
         // > Min을 통해 최종 최저비용 출력
         int oilPrice = Integer.MAX_VALUE;     // 지금까지 간 도시중 최저 오일 가격
         // (이후 방문한 국가의 오일보다 비쌀 경우, oilprice 가격으로 오일을 구매 (최저 oilPrice였던 국가에서 구매한걸로 침)
@@ -22,7 +22,7 @@ public class Main {
         for(int i=0; i<N-1; i++) {  // 한 국가씩 방문하면서 현재까지 나온 최저오일비용과 비교해가며 리터구매 (다음 국가까지 가는 거리 참고하여)
             oilPrice = Math.min(Integer.parseInt(oilPriceArr[i]), oilPrice);  // 방문국가의 오일가격과 이전에 지나왔던 최저오일비용과 비교하여 갱신
             // 현재 국가의 오일가격이 더 비싸면 최저오일비용으로 구매 (> 이전 국가에서 구매한걸로)
-            price += oilPrice * Integer.parseInt(distanceArr[i]);
+            price += oilPrice * Long.parseLong(distanceArr[i]);
         }
 
         bw.write(String.valueOf(price));
